Implemented add vehicles menu.

+++ Garage.Biz/Vehicles/Boat.cs
@@ -5,11 +5,11 @@ namespace Garage.Biz.Vehicles
     [Serializable]
     public class Boat : Vehicle
     {
-        private int length;
+        private double length;
 
-        public Boat(string regNr, string color, int wheelsNr, int length) : base(regNr, color, wheelsNr) => Length = length;
+        public Boat(string regNr, string color, int wheelsNr, double length) : base(regNr, color, wheelsNr) => Length = length;
 
-        public int Length { get => length; set => length = value; }
+        public double Length { get => length; set => length = value; }
 
         public override string ToString()
         {
